pyproject.toml
[project]
name = "craft-api-utilities"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.0",
    "duckdb>=1.4.1",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.2.250926",
    "pyarrow>=21.0.0",
    "pyarrow-stubs>=20.0.0.20250928",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
]
